added error boundary
