+++ templates/demo-store/app/routes/($lang)/$shopid/orders/$token/authenticate.tsx
@@ -1,0 +1,33 @@
+import {Shop} from '@shopify/hydrogen/storefront-api-types';
+import {redirect, type LoaderArgs} from '@shopify/remix-oxygen';
+import invariant from 'tiny-invariant';
+import {Button, PageHeader} from '~/components';
+
+export async function loader({request, context: {storefront}}: LoaderArgs) {
+  const {origin} = new URL(request.url);
+  const {shop} = await storefront.query<{
+    shop: Shop;
+  }>(`query getShopPrimaryDomain { shop { primaryDomain{ url } } }`, {
+    cache: storefront.CacheLong(),
+  });
+  invariant(shop, 'Error redirecting to the order status URL');
+  return redirect(request.url.replace(origin, shop.primaryDomain.url));
+}
+
+export default function () {
+  return null;
+}
+export function ErrorBoundary() {
+  return (
+    <PageHeader
+      heading={'Error redirecting to the order status URL'}
+      className="text-red-600"
+    >
+      <div className="flex items-baseline justify-between w-full">
+        <Button as="button" onClick={() => window.location.reload()}>
+          Try Again
+        </Button>
+      </div>
+    </PageHeader>
+  );
+}
